RMM doesn't try and find spdlog when it is embedded inside RMM
The conversion over to `rapids-cmake` missed an edge case. When
RMM builds spdlog itself it provided the `rmm::spdlog_header_only`
target which should be used instead of searching for `spdlog`.

This resolves issues where `find_package(rmm)` would ignore the fact it built `spdlog`, and instead find a build of `spdlog` and fails. This happens as the `spdlog` build directory has a broken `spdlog-config.cmake`.